refactor(android): isolate UI tree capture helpers
